requirements.txt
streamlit
fastapi
uvicorn[standard]
requests
aiohttp
beautifulsoup4
pandas
numpy
matplotlib
apscheduler
sqlalchemy
aiosqlite
schedule
python-dotenv
openai
yt-dlp
google-api-python-client
scikit-learn
tqdm
lxml
python-dateutil
plotly
python-multipart
